Move version number variables from base -> rest
