+++ base/git_info.go
@@ -1,4 +1,4 @@
-package rest
+package base
 
 // The git commit that was compiled. This will be filled in by the compiler.
 const GitProductName = ""
